+++ lib/stellar_client/client.rb
@@ -3,13 +3,14 @@ module StellarClient
 
     include APIClientBase::Client.module(default_opts: :default_opts)
     api_action :send_payment
+    api_action :get_toml
 
     attribute :bridge_host, String
 
     private
 
     def default_opts
-      { bridge_host: bridge_host }
+      { host: host, bridge_host: bridge_host }
     end
 
   end