+++ lib/stellar_client/requests/get_toml_request.rb
@@ -1,0 +1,11 @@
+module StellarClient
+  class GetTomlRequest
+
+    include APIClientBase::Request.module
+
+    def path
+      "/.well-known/stellar.toml"
+    end
+
+  end
+end