+++ lib/stellar_client/responses/get_toml_response.rb
@@ -1,0 +1,13 @@
+module StellarClient
+  class GetTomlResponse < BaseResponse
+
+    attribute :toml, Hash, lazy: true, default: :default_toml
+
+    private
+
+    def default_toml
+      Tomlrb.parse(raw_response.body)
+    end
+
+  end
+end