+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Renamed from BridgeClient to StellarClient
 - `host` configuration is now `bridge_host`
 
+### Added
+- `#get_toml` to fetch the toml of `host`
+
 ## [0.4.0] - 2017-04-05
 ### Added
 - Expose response's `code` (as `code_str`) and `message`